Fix sonar bugs

+++ src/creatures/players/vocations/vocation.h
@@ -60,19 +60,19 @@ class Vocation
 		}
 
 		uint32_t getManaGainTicks() const {
-			return gainManaTicks / g_configManager().getFloat(RATE_MANA_REGEN_SPEED);
+			return gainManaTicks / static_cast<uint32_t>(g_configManager().getFloat(RATE_MANA_REGEN_SPEED));
 		}
 
 		uint32_t getManaGainAmount() const {
-			return gainManaAmount * g_configManager().getFloat(RATE_MANA_REGEN);
+			return gainManaAmount * static_cast<uint32_t>(g_configManager().getFloat(RATE_MANA_REGEN));
 		}
 
 		uint32_t getHealthGainTicks() const {
-			return gainHealthTicks / g_configManager().getFloat(RATE_HEALTH_REGEN_SPEED);
+			return gainHealthTicks / static_cast<uint32_t>(g_configManager().getFloat(RATE_HEALTH_REGEN_SPEED));
 		}
 
 		uint32_t getHealthGainAmount() const {
-			return gainHealthAmount * g_configManager().getFloat(RATE_HEALTH_REGEN);
+			return gainHealthAmount * static_cast<uint32_t>(g_configManager().getFloat(RATE_HEALTH_REGEN));
 		}
 
 		uint8_t getSoulMax() const {
@@ -80,7 +80,7 @@ class Vocation
 		}
 
 		uint32_t getSoulGainTicks() const {
-			return gainSoulTicks / g_configManager().getFloat(RATE_SOUL_REGEN_SPEED);
+			return gainSoulTicks / static_cast<uint32_t>(g_configManager().getFloat(RATE_SOUL_REGEN_SPEED));
 		}
 
 		uint32_t getBaseAttackSpeed() const {
@@ -88,7 +88,7 @@ class Vocation
 		}
 
 		uint32_t getAttackSpeed() const {
-			return attackSpeed / g_configManager().getFloat(RATE_ATTACK_SPEED);
+			return attackSpeed / static_cast<uint32_t>(g_configManager().getFloat(RATE_ATTACK_SPEED));
 		}
 
 		uint32_t getBaseSpeed() const {
